Remove utils and put in lib
