+++ app/(protected)/(admin)/users/columns.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { ColumnHeader } from '@/components/column-header'
-import { PLACEHOLDER_AVATAR } from '@/utils/constants'
+import { PLACEHOLDER_AVATAR } from '@/lib/constants'
 import { Prisma } from '@prisma/client'
 import { ColumnDef } from '@tanstack/react-table'
 import Image from 'next/image'
