+++ app/(protected)/listings/[listingId]/edit/cards/listing-images.tsx
@@ -1,7 +1,7 @@
 import { deleteImage } from '@/actions/listing'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { PLACEHOLDER_IMAGE } from '@/utils/constants'
-import { getImageUrl } from '@/utils/helpers'
+import { PLACEHOLDER_IMAGE } from '@/lib/constants'
+import { getImageUrl } from '@/lib/utils'
 import { Prisma, ListingImage as TListingImage } from '@prisma/client'
 import Image from 'next/image'
 import UploadImages from '../components/upload-images'
