+++ app/(protected)/listings/columns.tsx
@@ -3,8 +3,8 @@
 import { ColumnHeader } from '@/components/column-header'
 import { Button } from '@/components/ui/button'
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuTrigger } from '@/components/ui/dropdown-menu'
-import { PLACEHOLDER_IMAGE } from '@/utils/constants'
-import { getImageUrl } from '@/utils/helpers'
+import { PLACEHOLDER_IMAGE } from '@/lib/constants'
+import { getImageUrl } from '@/lib/utils'
 import { createClient } from '@/utils/supabase/client'
 import { Listing, Prisma } from '@prisma/client'
 import { ColumnDef } from '@tanstack/react-table'
