+++ app/(protected)/settings/forms/avatar-form.tsx
@@ -2,8 +2,8 @@ import { updateAvatar } from '@/actions/user'
 import { ActionButton } from '@/components/action-button'
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
 import { Input } from '@/components/ui/input'
+import { PLACEHOLDER_AVATAR } from '@/lib/constants'
 import { AuthUser } from '@/lib/utils'
-import { PLACEHOLDER_AVATAR } from '@/utils/constants'
 import Image from 'next/image'
 
 export default function AvatarForm({ user }: { user: AuthUser }) {
