--- utils/supabase/client.ts
@@ -1,6 +1,0 @@
-import { Database } from '@/db_types'
-import { createBrowserClient } from '@supabase/ssr'
-
-export function createClient() {
-	return createBrowserClient<Database>(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!)
-}
